+++ xLights/models/Model.cpp
@@ -4558,7 +4558,7 @@ int Model::GetNodeStringNumber(size_t nodenum) const
 
 void Model::GetNode3DScreenCoords(int nodeidx, std::vector<std::tuple<float, float, float>>& pts)
 {
-    for (int x = 0; x < Nodes[nodeidx]->Coords.size(); ++x) {
+    for (size_t x = 0; x < Nodes[nodeidx]->Coords.size(); ++x) {
         pts.push_back(std::make_tuple(Nodes[nodeidx]->Coords[x].screenX, Nodes[nodeidx]->Coords[x].screenY, Nodes[nodeidx]->Coords[x].screenZ));
     }
 }
@@ -4567,7 +4567,7 @@ void Model::GetNodeCoords(int nodeidx, std::vector<wxPoint>& pts)
 {
     if (nodeidx >= Nodes.size())
         return;
-    for (int x = 0; x < Nodes[nodeidx]->Coords.size(); ++x) {
+    for (size_t x = 0; x < Nodes[nodeidx]->Coords.size(); ++x) {
         pts.push_back(wxPoint(Nodes[nodeidx]->Coords[x].bufX, Nodes[nodeidx]->Coords[x].bufY));
     }
 }
